refactor(models): extract HF/FLM pull-job orchestration to registry/pull_jobs
P3-routers §J: move the pull-job orchestration that was inlined in the
routes/models.py handlers — HF source/capability resolution, add-by-HF
registry seeding, the detached-task scheduler, the run_pull progress-event
wrapper, terminal footer emission, the persisted-snapshot disk fallback +
reconcile, and the FLM background pull — into a new deep module
hal0.registry.pull_jobs with a documented interface contract.

The route handlers (pull_model, update_model_from_hf, pull_status,
pull_stream, pull_cancel) stay as request→service→envelope shells. Route
paths, methods, payload shapes and status codes are unchanged.

Back-compat: routes.models keeps underscore-named re-export bindings
(_run_pull_with_events, _schedule_pull_task, _resolve_pull_capability, …)
so the monkeypatch seams in the test-suite and the hal0.api.__init__
imports of _run_pull_with_events/_schedule_pull_task keep working without
touching those files.

models.py: 1774 → 1298 lines.

test-import move: tests/api/test_models_pull_capability.py monkeypatches
get_curated where the moved helper now resolves it (hal0.registry.pull_jobs).
